+++ src/lib/presets/vanilla.js
@@ -2,13 +2,12 @@ import defaultStyles from '../default.css';
 import List from '../List.svelte';
 import Item from '../Item.svelte';
 import Selection from '../Selection.svelte';
-import Multi from '../Multi.svelte';
-
+import debounce from '../debounce';
 
 export default {
     theme: defaultStyles,
     List,
     Item,
     Selection,
-    Multi
+    debounce
 }